services/analytics-service/pyproject.toml
[project]
name = "analytics-service"
version = "0.1.0"
description = "analytics-service service"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.116,<1",
  "uvicorn[standard]>=0.35,<1",
  "pydantic>=2.8,<3",
  "nats-py>=2.8,<3",
  "prometheus-client>=0.23,<1"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
